requirements.txt
opencv-python>=4.8.0
PyQt5>=5.15.0
numpy>=1.24.0
mss>=9.0.0
PyYAML>=6.0

# ML / Detection
ultralytics>=8.0.0
torch>=2.0.0
roboflow>=1.0.0

# Tracking
filterpy>=1.4.5
scipy>=1.10.0

# Augmentation
albumentations>=1.3.0

# Hyperparameter optimization
optuna>=3.0.0

# Evaluation / visualization
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
